Command and query related to services were separated and Custom Exceptions handled.

+++ src/main/java/com/example/employee/employee_management/service/DepartmentCommandService.java
@@ -1,47 +1,41 @@
 package com.example.employee.employee_management.service;
 
+import com.example.employee.employee_management.exception.DuplicateEntityException;
+import com.example.employee.employee_management.exception.EntityNotFoundException;
 import com.example.employee.employee_management.model.Department;
 import com.example.employee.employee_management.repository.DepartmentRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 @Service
-
-public class DepartmentService {
+@Transactional
+public class DepartmentCommandService {
 
     DepartmentRepository departmentRepository;
 
-    public DepartmentService(DepartmentRepository departmentRepository) {
+    public DepartmentCommandService(DepartmentRepository departmentRepository) {
         this.departmentRepository = departmentRepository;
     }
 
-    public Optional<Department> findDepartmentById(Long id){
-        return departmentRepository.findById(id);
-    }
-
-    @Transactional
     public Department createDepartment(Department department){
+        if (departmentRepository.existsByName(department.getName())){
+            throw new DuplicateEntityException("Department with name '" + department.getName() + "' already exists");
+        }
         return departmentRepository.save(department);
     }
 
-    @Transactional
     public Department updateDepartment(Long id, Department updatedDepartment){
         return departmentRepository.findById(id)
                 .map(department -> {
                     department.setName(updatedDepartment.getName());
                     return departmentRepository.save(department);
-                }).orElseThrow(() -> new RuntimeException("Department not found with id: " + id));
+                }).orElseThrow(() -> new EntityNotFoundException("Department", id));
     }
 
-    @Transactional
     public void deleteDepartmentById(Long id){
+        if (!departmentRepository.existsById(id)){
+            throw new EntityNotFoundException("Department", id);
+        }
         departmentRepository.deleteById(id);
-    }
-
-    @Transactional
-    public Optional<Department> findDepartmentByIdWithEmployees(Long id){
-        return departmentRepository.findDepartmentByIdWithEmployees(id);
     }
 }
